+++ src/settings/NumberGeneratorSettings/constants.js
@@ -1,0 +1,20 @@
+export const NUMBER_GENERATOR_OPTIONS_OFF = 'off';
+export const NUMBER_GENERATOR_OPTIONS_ON_EDITABLE = 'onEditable';
+export const NUMBER_GENERATOR_OPTIONS_ON_NOT_EDITABLE = 'onNotEditable';
+
+export const NUMBER_GENERATOR_OPTIONS = [
+  { value: '' },
+  { value: NUMBER_GENERATOR_OPTIONS_OFF },
+  { value: NUMBER_GENERATOR_OPTIONS_ON_EDITABLE },
+  { value: NUMBER_GENERATOR_OPTIONS_ON_NOT_EDITABLE },
+];
+
+export const NUMBER_GENERATOR_SETTINGS_KEY = 'number-generator-settings';
+export const NUMBER_GENERATOR_SETTINGS_SCOPE = 'ui-users.number-generator-settings.manage';
+
+export const BARCODE_SETTING = 'barcode';
+
+export const SERVICE_INTERACTION_API = '/settings/service-interaction';
+export const SERVICE_INTERACTION_NUMBER_GENERATOR_SEQUENCES_API = '/settings/service-interaction/numberGeneratorSequences/';
+
+export const MOD_SETTINGS_API = 'settings/entries';